fix: stabilize attestation state and adaptive defaults
