--- .github/scripts/apply_attestation_state_hotfix.py
@@ -1,275 +1,0 @@
-from pathlib import Path
-import re
-
-
-def sub1(path, pattern, replacement, flags=0):
-    p = Path(path)
-    text = p.read_text()
-    new, count = re.subn(pattern, replacement, text, count=1, flags=flags)
-    if count != 1:
-        raise SystemExit(f"expected one regex match in {path}, got {count}: {pattern[:100]}")
-    p.write_text(new)
-
-
-policy = "service/src/main/java/cleveres/tricky/cleverestech/PolicyState.kt"
-sub1(
-    policy,
-    r'val parsed = parseStateFile\(main, "configured"\)\n\s+publish\(parsed, persistPrevious = false\)',
-    'val parsed = adaptRecommendedDefault(parseStateFile(main, "configured"))\n            publish(parsed, persistPrevious = false)',
-)
-sub1(
-    policy,
-    r'    @Synchronized\n    fun applyRecommendedDefaults\(\) \{.*?\n    \}\n\n    @Synchronized\n    fun synchronizeBuiltInProfile',
-    '''    private fun recommendedSecurityPatchEnabled(thresholdMonths: Long = 6): Boolean {
-        val now = currentDateSource()
-        val available = listOf("system", "vendor", "boot").mapNotNull(::readPropertyDate)
-        return available.any { it.isBefore(now.minusMonths(thresholdMonths)) }
-    }
-
-    private fun isRecommendedDefaultShape(value: Snapshot): Boolean {
-        val features = value.features
-        val patch = value.patch
-        return value.profiles.isEmpty() && value.activeProfile == null &&
-            !features.buildIdentity && !features.attestationIdentity && !features.telephonyIdentity &&
-            !features.regionIdentity && !features.identityRefresh &&
-            patch.thresholdMonths == 6L &&
-            patch.system.mode == PatchMode.AUTOMATIC &&
-            patch.vendor.mode == PatchMode.AUTOMATIC &&
-            patch.boot.mode == PatchMode.AUTOMATIC
-    }
-
-    private fun adaptRecommendedDefault(value: Snapshot): Snapshot {
-        if (!isRecommendedDefaultShape(value)) return value
-        val enabled = recommendedSecurityPatchEnabled(value.patch.thresholdMonths)
-        if (value.features.securityPatch == enabled) return value
-        return value.copy(
-            features = value.features.copy(securityPatch = enabled),
-            generation = generationCounter.incrementAndGet(),
-            recovery = "default_adapted",
-        )
-    }
-
-    @Synchronized
-    fun applyRecommendedDefaults() {
-        val automatic = PatchPolicy(PatchMode.AUTOMATIC)
-        val enabled = recommendedSecurityPatchEnabled(6)
-        persistAndPublish(
-            Snapshot(
-                explicit = true,
-                features = FeatureSet(false, false, false, false, false, enabled),
-                patch = PatchSet(6, automatic, automatic, automatic),
-                profiles = emptyMap(),
-                activeProfile = null,
-                assignments = emptyList(),
-                generation = generationCounter.incrementAndGet(),
-                recovery = "default",
-            ),
-        )
-    }
-
-    @Synchronized
-    fun synchronizeBuiltInProfile''',
-    re.S,
-)
-
-cert = "service/src/main/java/cleveres/tricky/cleverestech/keystore/CertHack.java"
-sub1(
-    cert,
-    r'        final Map<KeyBox, PreparedKeyBox> preparedKeyboxes;\n        final Map<CacheKey, Certificate\[]> certificateCache;',
-    '''        final Map<KeyBox, PreparedKeyBox> preparedKeyboxes;
-        final byte[] bootKey;
-        final byte[] bootHash;
-        final Map<CacheKey, Certificate[]> certificateCache;''',
-)
-sub1(
-    cert,
-    r'        State\(Map<String, List<KeyBox>> keyboxes, Map<String, List<KeyBox>> keyboxFiles\) \{.*?\n        \}\n\n        private static Map<String, List<KeyBox>> immutableLists',
-    '''        State(Map<String, List<KeyBox>> keyboxes, Map<String, List<KeyBox>> keyboxFiles) {
-            IdentityHashMap<KeyBox, PreparedKeyBox> prepared = new IdentityHashMap<>();
-            Map<String, List<KeyBox>> readyByAlgorithm = new HashMap<>();
-            Map<String, List<KeyBox>> readyByFile = new HashMap<>();
-            for (Map.Entry<String, List<KeyBox>> entry : keyboxes.entrySet()) {
-                for (KeyBox keybox : entry.getValue()) {
-                    if (prepared.containsKey(keybox)) continue;
-                    try {
-                        prepared.put(keybox, new PreparedKeyBox(keybox));
-                        readyByAlgorithm.computeIfAbsent(entry.getKey(), ignored -> new ArrayList<>()).add(keybox);
-                        readyByFile.computeIfAbsent(keybox.filename, ignored -> new ArrayList<>()).add(keybox);
-                    } catch (Exception error) {
-                        Logger.e("Ignoring an unprepared keybox", error);
-                    }
-                }
-            }
-            this.keyboxes = immutableLists(readyByAlgorithm);
-            this.keyboxFiles = immutableLists(readyByFile);
-            this.preparedKeyboxes = Collections.unmodifiableMap(prepared);
-            if (prepared.isEmpty()) {
-                this.bootKey = null;
-                this.bootHash = null;
-            } else {
-                byte[] resolvedBootKey = usableBootDigest(UtilKt.getBootKey());
-                byte[] resolvedBootHash = usableBootDigest(UtilKt.getBootHash());
-                if (resolvedBootKey == null) resolvedBootKey = usableBootDigest(UtilKt.getPersistentBootKey());
-                if (resolvedBootHash == null) resolvedBootHash = usableBootDigest(UtilKt.getPersistentBootHash());
-                this.bootKey = resolvedBootKey == null ? null : resolvedBootKey.clone();
-                this.bootHash = resolvedBootHash == null ? null : resolvedBootHash.clone();
-            }
-            this.certificateCache = Collections.synchronizedMap(
-                    new LinkedHashMap<CacheKey, Certificate[]>(32, 0.75f, true) {
-                        @Override
-                        protected boolean removeEldestEntry(Map.Entry<CacheKey, Certificate[]> eldest) {
-                            return size() > MAX_CERTIFICATE_CACHE_ENTRIES;
-                        }
-                    });
-        }
-
-        private static Map<String, List<KeyBox>> immutableLists''',
-    re.S,
-)
-sub1(
-    cert,
-    r'    public static boolean canHack\(\) \{\n        return !state\.keyboxes\.isEmpty\(\);\n    \}',
-    '''    public static boolean canHack() {
-        State currentState = state;
-        return !currentState.keyboxes.isEmpty() && currentState.bootKey != null && currentState.bootHash != null;
-    }''',
-)
-sub1(
-    cert,
-    r'            byte\[] verifiedBootKey = usableBootDigest\(UtilKt\.getBootKey\(\)\);\n            byte\[] verifiedBootHash = usableBootDigest\(UtilKt\.getBootHash\(\)\);',
-    '''            byte[] verifiedBootKey = currentState.bootKey == null ? null : currentState.bootKey.clone();
-            byte[] verifiedBootHash = currentState.bootHash == null ? null : currentState.bootHash.clone();''',
-)
-sub1(
-    cert,
-    r'            ASN1Encodable\[] rootOfTrustEnc = \{.*?            ASN1Sequence hackedRootOfTrust = new DERSequence\(rootOfTrustEnc\);',
-    '            ASN1Sequence hackedRootOfTrust = buildLockedRootOfTrust(verifiedBootKey, verifiedBootHash);',
-    re.S,
-)
-sub1(
-    cert,
-    r'    private static byte\[] usableBootDigest\(byte\[] value\) \{',
-    '''    static ASN1Sequence buildLockedRootOfTrust(byte[] verifiedBootKey, byte[] verifiedBootHash) {
-        Objects.requireNonNull(verifiedBootKey, "verifiedBootKey");
-        Objects.requireNonNull(verifiedBootHash, "verifiedBootHash");
-        ASN1Encodable[] rootOfTrustEnc = {
-                new DEROctetString(verifiedBootKey),
-                ASN1Boolean.TRUE,
-                new ASN1Enumerated(0),
-                new DEROctetString(verifiedBootHash)
-        };
-        return new DERSequence(rootOfTrustEnc);
-    }
-
-    private static byte[] usableBootDigest(byte[] value) {''',
-)
-
-template = Path("module/template/policy_state_v2.json")
-raw = template.read_text()
-if '"securityPatch": true' not in raw:
-    raise SystemExit("unexpected default policy template")
-template.write_text(raw.replace('"securityPatch": true', '"securityPatch": false', 1))
-
-test = "service/src/test/java/cleveres/tricky/cleverestech/PolicyStateHotPathTest.kt"
-sub1(
-    test,
-    r'    @Test\n    fun recommendedDefaultsStayAligned\(\) \{.*?\n    \}\n\n    @Test\n    fun capturedAutomaticPatchSkipsPropertyReads',
-    '''    @Test
-    fun currentDeviceKeepsAutomaticSecurityPatchDisabledByDefault() {
-        val root = Files.createTempDirectory("ct-defaults-current").toFile()
-        val oldProperties = systemPropertiesGet
-        try {
-            PolicyState.setRootForTesting(root)
-            systemPropertiesGet = { key, default ->
-                when (key) {
-                    "ro.build.version.security_patch",
-                    "ro.vendor.build.security_patch",
-                    "ro.bootimage.build.version.security_patch" -> "2026-08-05"
-                    else -> default
-                }
-            }
-            PolicyState.currentDateSource = { LocalDate.of(2026, 8, 14) }
-            PolicyState.applyRecommendedDefaults()
-            val state = PolicyState.stateJson()
-            val features = state.getJSONObject("features")
-            assertFalse(features.getBoolean("buildIdentity"))
-            assertFalse(features.getBoolean("attestationIdentity"))
-            assertFalse(features.getBoolean("telephonyIdentity"))
-            assertFalse(features.getBoolean("regionIdentity"))
-            assertFalse(features.getBoolean("identityRefresh"))
-            assertFalse(features.getBoolean("securityPatch"))
-            val patch = state.getJSONObject("securityPatch")
-            assertEquals(6L, patch.getLong("automaticThresholdMonths"))
-            listOf("system", "vendor", "boot").forEach { component ->
-                assertEquals("automatic", patch.getJSONObject(component).getString("mode"))
-            }
-        } finally {
-            systemPropertiesGet = oldProperties
-            PolicyState.resetForTesting()
-            root.deleteRecursively()
-        }
-    }
-
-    @Test
-    fun staleDeviceEnablesAutomaticSecurityPatchByDefault() {
-        val root = Files.createTempDirectory("ct-defaults-stale").toFile()
-        val oldProperties = systemPropertiesGet
-        try {
-            PolicyState.setRootForTesting(root)
-            systemPropertiesGet = { key, default ->
-                when (key) {
-                    "ro.build.version.security_patch" -> "2025-01-05"
-                    "ro.vendor.build.security_patch",
-                    "ro.bootimage.build.version.security_patch" -> "2026-08-05"
-                    else -> default
-                }
-            }
-            PolicyState.currentDateSource = { LocalDate.of(2026, 8, 14) }
-            PolicyState.applyRecommendedDefaults()
-            assertTrue(PolicyState.stateJson().getJSONObject("features").getBoolean("securityPatch"))
-        } finally {
-            systemPropertiesGet = oldProperties
-            PolicyState.resetForTesting()
-            root.deleteRecursively()
-        }
-    }
-
-    @Test
-    fun capturedAutomaticPatchSkipsPropertyReads''',
-    re.S,
-)
-
-certtest = "service/src/test/java/cleveres/tricky/cleverestech/keystore/CertHackTest.java"
-sub1(
-    certtest,
-    r'import org\.junit\.After;',
-    '''import org.bouncycastle.asn1.ASN1Boolean;
-import org.bouncycastle.asn1.ASN1Enumerated;
-import org.bouncycastle.asn1.ASN1OctetString;
-import org.bouncycastle.asn1.ASN1Sequence;
-import org.junit.After;''',
-)
-sub1(
-    certtest,
-    r'\n    @Test\n    public void testSigningKeyAlgorithmUsesCertificateSigner\(\)',
-    '''
-    @Test
-    public void testLockedRootOfTrustIsStable() {
-        byte[] key = new byte[32];
-        byte[] hash = new byte[32];
-        key[0] = 1;
-        hash[0] = 2;
-        ASN1Sequence root = CertHack.buildLockedRootOfTrust(key, hash);
-        assertArrayEquals(key, ASN1OctetString.getInstance(root.getObjectAt(0)).getOctets());
-        assertTrue(ASN1Boolean.getInstance(root.getObjectAt(1)).isTrue());
-        assertEquals(0, ASN1Enumerated.getInstance(root.getObjectAt(2)).getValue().intValue());
-        assertArrayEquals(hash, ASN1OctetString.getInstance(root.getObjectAt(3)).getOctets());
-    }
-
-    @Test
-    public void testSigningKeyAlgorithmUsesCertificateSigner()''',
-)
-
-for path in Path('.github/workflows').glob('attestation-state-hotfix*.yml'):
-    path.unlink()
-Path('.github/scripts/apply_attestation_state_hotfix.py').unlink()
