--- .github/workflows/attestation-state-hotfix-retry.yml
@@ -1,61 +1,0 @@
-name: Attestation state hotfix retry
-
-on:
-  push:
-    branches:
-      - fix/v2.5.7-attestation-state-defaults
-
-permissions:
-  contents: write
-
-jobs:
-  patch:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1
-        with:
-          fetch-depth: 0
-
-      - name: Setup Java 21
-        uses: actions/setup-java@b6effb05e454b25005698d916606bdc6ffcbf961
-        with:
-          distribution: temurin
-          java-version: 21
-
-      - name: Replay patch with normalized blocks
-        shell: bash
-        run: |
-          set -euo pipefail
-          python3 <<'PY'
-          from pathlib import Path
-          source = Path('.github/workflows/attestation-state-hotfix.yml').read_text()
-          start = source.index("          python3 <<'PY'\n") + len("          python3 <<'PY'\n")
-          end = source.index("          PY\n", start)
-          lines = source[start:end].splitlines()
-          script = '\n'.join(line[10:] if line.startswith('          ') else line for line in lines) + '\n'
-          script = script.replace('from pathlib import Path\n', 'from pathlib import Path\nimport textwrap\n', 1)
-          script = script.replace('def replace_once(path, old, new):\n    p = Path(path)\n', 'def replace_once(path, old, new):\n    old = textwrap.dedent(old)\n    new = textwrap.dedent(new)\n    p = Path(path)\n', 1)
-          Path('/tmp/apply.py').write_text(script)
-          PY
-          python3 /tmp/apply.py
-          rm -f .github/workflows/attestation-state-hotfix-retry.yml
-
-      - name: Validate focused changes
-        shell: bash
-        run: |
-          set -euo pipefail
-          ./gradlew :service:testDebugUnitTest --no-daemon
-          for test in module/webui-tests/*.test.js; do node "$test"; done
-          git diff --check
-
-      - name: Commit hotfix
-        shell: bash
-        run: |
-          set -euo pipefail
-          git config user.name "Yiğit"
-          git config user.email "40565628+tryigit@users.noreply.github.com"
-          git add -A
-          git diff --cached --check
-          git commit -m "fix: stabilize attestation state and adaptive defaults"
-          git push origin HEAD
