--- .github/workflows/run-attestation-state-hotfix.yml
@@ -1,31 +1,0 @@
-name: Run attestation state hotfix
-on:
-  push:
-    branches: [fix/v2.5.7-attestation-state-defaults]
-permissions:
-  contents: write
-jobs:
-  patch:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1
-        with:
-          fetch-depth: 0
-      - uses: actions/setup-java@b6effb05e454b25005698d916606bdc6ffcbf961
-        with:
-          distribution: temurin
-          java-version: 21
-      - run: python3 .github/scripts/apply_attestation_state_hotfix.py
-      - run: python3 .github/scripts/fix_default_test.py
-      - run: ./gradlew :service:testDebugUnitTest --no-daemon
-      - run: |
-          for test in module/webui-tests/*.test.js; do node "$test"; done
-          git diff --check
-      - run: |
-          rm -f .github/workflows/run-attestation-state-hotfix.yml
-          git config user.name "Yiğit"
-          git config user.email "40565628+tryigit@users.noreply.github.com"
-          git add -A
-          git diff --cached --check
-          git commit -m "fix: stabilize attestation state and adaptive defaults"
-          git push origin HEAD
