+++ module/template/policy_state_v2.json
@@ -6,7 +6,7 @@
     "telephonyIdentity": false,
     "regionIdentity": false,
     "identityRefresh": false,
-    "securityPatch": true
+    "securityPatch": false
   },
   "securityPatch": {
     "automaticThresholdMonths": 6,
